fix(firebaseServices): Change names of all the exam pages to ending with either Identify or Notate, and also change corresponding variables in each file and the home page view states.
